requirements.txt
mcp>=1.0.0
voyageai>=0.2.3
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
